Add happening
Merge event and bedpres into happening. Fix sanity importer. Update some schemas
