+++ apps/cms/schemas/index.ts
@@ -1,12 +1,10 @@
-import bedpres from "./bedpres";
 import company from "./company";
-import event from "./event";
+import happening from "./happening";
 import job from "./job-ad";
 import location from "./location";
 import meetingMinute from "./meeting-minute";
 import contactProfile from "./objects/contact-profile";
 import question from "./objects/question";
-import registrationDates from "./objects/registration-dates";
 import spotRange from "./objects/spot-range";
 import post from "./post";
 import profile from "./profile";
@@ -14,8 +12,7 @@ import staticInfo from "./static-info";
 import studentgroup from "./student-group";
 
 export const schemaTypes = [
-  event,
-  bedpres,
+  happening,
   post,
   company,
   job,
@@ -27,5 +24,4 @@ export const schemaTypes = [
   spotRange,
   contactProfile,
   question,
-  registrationDates,
 ];
